# core/src/server/openapi.yaml
openapi: 3.0.0
info:
  title: PMXT Sidecar API
  description: >-
    A unified local sidecar API for prediction markets (Polymarket, Kalshi, Limitless). This API acts as a
    JSON-RPC-style gateway. Each endpoint corresponds to a specific method on the generic exchange implementation.
  version: 0.4.4
servers:
  - url: 'http://localhost:3847'
    description: Local development server
paths:
...
  '/api/{exchange}/watchOrderBook':
    post:
      summary: Watch Order Book
      operationId: watchOrderBook
      parameters:
        - in: path
          name: exchange
          schema:
            type: string
            enum:
              - polymarket
              - kalshi
              - kalshi-demo
              - limitless
              - probable
              - baozi
              - myriad
              - opinion
              - polymarket_us
          required: true
          description: The prediction market exchange to target.
      requestBody:
        content:
          application/json:
            schema:
              title: WatchOrderBookRequest
              type: object
              properties:
                args:
                  type: array
                  minItems: 1
                  maxItems: 2
                  items:
                    oneOf:
                      - type: string
                      - type: number
                credentials:
                  $ref: '#/components/schemas/ExchangeCredentials'
              required:
                - args
      responses:
        '200':
          description: Watch Order Book response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderBook'
      description: >-
        Watch order book updates in real-time via WebSocket. Returns a promise that resolves with the next order book
        update. Call repeatedly in a loop to stream updates (CCXT Pro pattern).
  '/api/{exchange}/watchOrderBooks':
    post:
      summary: Watch Order Books
      operationId: watchOrderBooks
      parameters:
        - in: path
          name: exchange
          schema:
            type: string
            enum:
              - polymarket
              - kalshi
              - kalshi-demo
              - limitless
              - probable
              - baozi
              - myriad
              - opinion
              - polymarket_us
          required: true
          description: The prediction market exchange to target.
      requestBody:
        content:
          application/json:
            schema:
              title: WatchOrderBooksRequest
              type: object
              properties:
                args:
                  type: array
                  minItems: 1
                  maxItems: 2
                  items:
                    oneOf:
                      - type: array
                        items:
                          type: string
                      - type: number
                credentials:
                  $ref: '#/components/schemas/ExchangeCredentials'
              required:
                - args
      responses:
        '200':
          description: Watch Order Books response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties:
                          $ref: '#/components/schemas/OrderBook'
      description: >-
        Watch multiple order books simultaneously via WebSocket. Returns a promise that resolves with a record of order
        book snapshots keyed by ID. Exchanges with native batch support (e.g. Kalshi) send a single subscribe message
        for all tickers; others fall back to individual watchOrderBook calls.
...
  '/api/{exchange}/watchTrades':
    post:
      summary: Watch Trades
      operationId: watchTrades
      parameters:
        - in: path
          name: exchange
          schema:
            type: string
            enum:
              - polymarket
              - kalshi
              - kalshi-demo
              - limitless
              - myriad
              - opinion
              - polymarket_us
          required: true
          description: The prediction market exchange to target.
      requestBody:
        content:
          application/json:
            schema:
              title: WatchTradesRequest
              type: object
              properties:
                args:
                  type: array
                  minItems: 1
                  maxItems: 4
                  items:
                    oneOf:
                      - type: string
                      - type: string
                      - type: number
                      - type: number
                credentials:
                  $ref: '#/components/schemas/ExchangeCredentials'
              required:
                - args
      responses:
        '200':
          description: Watch Trades response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Trade'
      description: >-
        Watch trade executions in real-time via WebSocket. Returns a promise that resolves with the next trade(s). Call
        repeatedly in a loop to stream updates (CCXT Pro pattern).
...
components:
...
  schemas:
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      properties:
        message:
          type: string
...
    OrderBook:
      type: object
      properties:
        bids:
          type: array
          items:
            $ref: '#/components/schemas/OrderLevel'
          description: 'Order book bid levels, sorted by price descending.'
        asks:
          type: array
          items:
            $ref: '#/components/schemas/OrderLevel'
          description: 'Order book ask levels, sorted by price ascending.'
        timestamp:
          type: number
          description: Unix timestamp in milliseconds when the snapshot was taken.
      required:
        - bids
        - asks
    OrderLevel:
      type: object
      properties:
        price:
          type: number
          description: 0.0 to 1.0 (probability)
        size:
          type: number
          description: contracts/shares
      required:
        - price
        - size
    Trade:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier for this trade.
        timestamp:
          type: number
          description: Unix timestamp in milliseconds when the trade executed.
        price:
          type: number
          description: Probability between 0.0 and 1.0.
        amount:
          type: number
          description: Size of the trade in contracts/shares.
        side:
          type: string
          enum:
            - buy
            - sell
            - unknown
          description: Trade side from the taker's perspective.
        outcomeId:
          type: string
          description: The outcome this trade is for (if known).
      required:
        - id
        - timestamp
        - price
        - amount
        - side
...
    ExchangeCredentials:
      type: object
      description: Optional authentication credentials for exchange operations.
      properties:
        apiKey:
          type: string
        apiSecret:
          type: string
          description: Standard API secret for HMAC-authenticated exchanges
        passphrase:
          type: string
          description: Standard API passphrase for HMAC-authenticated exchanges
        apiToken:
          type: string
          description: 'Metaculus: `Authorization: Token <apiToken>` for higher rate limits'
        privateKey:
          type: string
          description: Required for Polymarket L1 auth
        signatureType:
          oneOf:
            - type: number
            - type: string
          description: '0 = EOA, 1 = Poly Proxy, 2 = Gnosis Safe (Can also use ''eoa'', ''polyproxy'', ''gnosis_safe'')'
        funderAddress:
          type: string
          description: The address funding the trades (defaults to signer address)
        walletAddress:
          type: string
...